requirements.txt
aniso8601==8.0.0
astroid==2.3.0
autopep8==1.4.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-JWT-Extended==3.23.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.1
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
